+++ feedme/services/FeedMe_CacheService.php
@@ -3,6 +3,9 @@ namespace Craft;
 
 class FeedMe_CacheService extends BaseApplicationComponent
 {
+    // Public Methods
+    // =========================================================================
+
     public function set($url, $value, $duration)
     {
         return craft()->cache->set(base64_encode(urlencode($url)), $value, $duration, null);